requirement.txt
cx_Oracle
paramiko
mysql-connector